test(arangodb): migrate unit tests to internal/arangodb package
