+++ internal/arangodb/setup_test.go
@@ -1,4 +1,4 @@
-package main
+package arangodb
 
 import (
 	"context"
